Move App state to PostPage component
