--- instagram-app/src/App.css
@@ -1,4 +1,0 @@
-.container {
-  width: 80%;
-  margin: 0 auto;
-}
