+++ instagram-app/src/components/presentations/PostsPage/PostsPage.css
@@ -24,3 +24,9 @@
   flex-wrap: wrap;
   width: 650px;
 }
+
+.last-comment-date {
+  font-size: 1.5rem;
+  color: #beb9b9;
+  font-weight: 400;
+}
